REC2017Programming/2017rec/requirements.txt
aniso8601==1.3.0
click==6.7
enum-compat==0.0.2
eventlet==0.21.0
Flask==0.12.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.6
greenlet==0.4.12
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-dateutil==2.6.1
python-engineio==1.7.0
python-socketio==1.8.1
pytz==2017.3
six==1.11.0
Werkzeug==0.12.2
